Chapter 11/cloud/voteprocessor/requirements.txt
gunicorn==23.0.0
Flask==3.0.3
Werkzeug==3.0.3
google-cloud-pubsub
google-api-core
google-cloud-spanner